requirements.txt
coloredlogs==15.0.1
humanize==4.7.0
fastapi==0.99.1
fastapi-utils==0.2.1
python-decouple==3.8
pytz==2023.3
requests==2.31.0
uvicorn==0.22.0
verboselogs==1.7
wakeonlan==3.0.0
pyautogui==0.9.54